requirements.txt
altair==4.2.0
altgraph==0.17.3
annotated-types==0.6.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
aspose-words==23.8.0
astropy==5.3.1
asttokens==2.2.1
async-generator==1.10
attr==0.3.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.5
cachetools==5.2.0
camera==1.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.3
commonmark==0.9.1
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
distro==1.8.0
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
ez-setup==0.9
fastjsonschema==2.16.3
filelock==3.7.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
fonttools==4.42.0
fqdn==1.5.1
gitdb==4.0.9
GitPython==3.1.27
greenlet==2.0.1
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
imageio==2.31.1
importlib-metadata==5.0.0
ipykernel==6.22.0
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-nbextensions-configurator==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterthemes==0.20.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy_loader==0.3
lesscpy==0.15.1
lxml==4.9.3
MarkupSafe==2.1.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==2.0.5
morpheus-astro==0.5.1
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.23.5
openai==1.1.1
opencv-camera==0.11.0
opencv-python==4.6.0.66
outcome==1.2.0
packaging==21.3
pandas==2.0.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.2
ply==3.11
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==10.0.0
pycountry==22.3.5
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.0b4
pyerfa==2.0.0.3
Pygments==2.15.1
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
Pympler==1.0.1
pyparsing==3.0.9
pypng==0.20220715.0
PyQt6==6.5.2
PyQt6-Qt6==6.5.2
PyQt6-sip==13.5.2
pyrsistent==0.19.2
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytictoc==1.5.3
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==306
pywin32-ctypes==0.2.0
pywinpty==2.0.10
PyYAML==6.0
pyzbar==0.1.9
pyzmq==25.0.2
qrcode==7.4.2
qtconsole==5.4.3
QtPy==2.3.1
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
scikit-image==0.21.0
scipy==1.8.1
selenium==4.15.2
semver==2.13.0
Send2Trash==1.8.2
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==1.4.44
stack-data==0.6.2
statsmodels==0.13.2
streamlit==1.14.0
terminado==0.17.1
tifffile==2023.8.12
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.8.0
tzdata==2022.6
tzlocal==4.2
uri-template==1.2.0
urllib3==1.26.12
validators==0.20.0
virtualenv==20.14.1
watchdog==2.1.9
wcwidth==0.2.6
webcolors==1.13
webdriver-manager==3.8.5
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.7
wsproto==1.2.0
zipp==3.10.0
